Download files needed for selfie capture
When testing the new Acuant SDK v11.9.3, I found parts of the selfie
functionality didn't work.

After research, I realized that was because we had missing files. Our
`download_acuant_sdk` script only copied over `.min.js` and `.wasm`
files.

We found after talking with Matt Hinz that this was because previously
we only needed the document capture files.

With this change, we should get everything that's needed - there is
currently one file we aren't downloading - and that's the
`AcuantCamera.js`. There are two reasons I chose not to download that.

1. the same code is already in place and used in `AcuantCamera.min.js`.
2. it seems like it'd be easier to just download everything, but we do
want to be mindful of not including more files than we need.

There is a chance that this may come up again in a future upgrade, but
that is why I am updating our `docs/sdk-upgrade.md` as well.

Co-authored-by: Eileen McFarland <eileen.mcfarland@gsa.gov>