LG-12674 Ensure requested VTR is logged in SAML and OIDC endpoints
This commit adds logging to ensure we have visibility into the VTR param that is sent by a service provider using SAML or OIDC.

When the param is sent it is expanded to include implied components. For example, “Pb” will be expanded into “C1.C2.P1.Pb”. For debugging purposes we will want visibility into what the service provider actually requests.

Additionally, when looking at SP redirects that are served it may be helpful to see the VTR and ACR values that were operated on.

This addresses this concern on the following events:

- OpenID Connect: authorization request
- SAML Auth
- SAML Auth Request
- SP redirect initiated

This commit also updates the SAML request logging params in `analytics_events` to be inclusive of what is actually logged.

### OpenID Connect: authorization request

This request already has a `vtr` value that is logged. However this is the parsed VTR value. If the VTR cannot be parsed this value is nil. This commit adds a new `vtr_param` value that includes the raw, unparsed `vtr` param.

### SAML Auth

This event logs the result from `SamlRequestValidator#call` which includes the AuthnContext as an extra analytic attribute. This is where the raw VTR is read from. No changes were necessary here besides better documentation of the params.

### SAML Auth Request

This commit added logging of the AuthnContext here to include both the raw ACR values and VTR param that are requested.

### SP redirect initiated

This commit added logging of the `vtr` and `acr_values` that are present in `sp_session`. This will allow us to see the ACR and VTR values that were operated on.

[skip changelog]