LG-12365 Log rate limit status on verify-by-mail enter-code controller
We have serveral rate limits that are enforced when the user visits the verify-by-mail code entry screen:

1. _OTP Rate-Limit_: This limit applies to OTP entries. After the user has entered too many OTPs this is enforced and the user is redirected to an error screen.
2. _Letter request limit_: This limit applies to letter requests. After the user has requested too many letters or if the users profile is too old the user no longer sees the option to request a new letter.

The enter code screen enforced these limits but did not add any indicator to the logs that it was doing so. This commit adds properties to the analytics event for each of these cases so we can monitor how the app is behaving.

[skip changelog]