Add the ability to render an expanded `AccordionComponent`
In [LG-12064](https://cm-jira.usa.gov/browse/LG-12064) we have a change to render an accordion that is expanded by default on screen. Currently `AccordionComponent` does not support being rendered in this way.

The USWDS suggests that the underlying component can be [rendered open by default by setting the `aria-exapnded` attribute](https://designsystem.digital.gov/components/accordion/#using-the-accordion-component-2):

> Add the `aria-expanded="true"` attribute to any `usa-accordion__button` to have that section open by default at page load. When the accordion is initialized, the JavaScript will automatically add `aria-expanded="false"` attribute to all other accordion buttons.

This commit adds an `expanded` option to `AccordionComponent` and renders the component with `aria-expanded="true”` if it is set to true. The default value is false to maintain existing behavior.

[skip changelog]