Split separate verify-by-mail enter-code contexts into partials (#10327)
The `Idv::ByMail::EnterCodeController` is responsible for rendering a screen where the user can enter their code. It renders a form for the user to enter a code in 2 contexts:

1. When the user signs in with a pending GPO profile
2. When the user clicks on the “sign in to request another letter” link on the “Finish verifying your identity” reminder email

These contexts are conveyed to the view using the `@user_did_not_receive_letter` and several changes are made accordingly.

In [LG-11753](https://cm-jira.usa.gov/browse/LG-11753) and [LG-12064](https://cm-jira.usa.gov/browse/LG-12064) we are making dramatic changes to these screens. Both contexts look very different so this commit breaks them into partials so they can be done independently.

There is some shared code that is moved into the partials that can be moved out when this work is finished. Namely the alert that a user cannot send more letters and the “Return to your profile” link on the bottom of the page. I have opted to move everything into the partials so we can modify the screens separately and then break out shared content in a follow up.

[skip changelog]