Fix issue where the “Did not receive letter” UI appears after incorrect submission (#10341)
The verify-by-mail flow’s enter code screen has a hidden field for the “did_not_receive_letter” param. This allowed the “Did not receive letter” UI to be applied on failed submission by inspecting the param like so:

```ruby
def user_did_not_receive_letter?
  !!params[:did_not_receive_letter]
end
```

However, the hidden input submits an empty string when the `@user_did_not_receive_letter` ivar is set to false. The empty string is truthy which made the UI revert to the “Did not receive letter” experience.

This commit fixes the issue by calling `present?` on the param.

[skip changelog]