Test h1 and body copy logic for doc auth + selfie
This change:

- moves some test setup to before all / after all
  - this is so we can reuse the user and other vars we do not change in
  this file and do not care about
  - the one drawback is that currently if you ctrl + c twice out of this
  test, the `after(:all)` block will not run and you will end up with a
  user hanging out in the db that will not be deleted
    - we discussed fixing this in this test, but decided it wasn't a
    good time. If folks agree we can create a ticket to come back to
    this work
    - there is no issue if you let the test finish on its own, or if you
    only ctrl + c once, letting it clean up and run the `after` block
      - pressing it twice skips the `after` block, which is how it ends
      up in this broken state
- adds a lot of test fixture files
  - we needed a lot of different setup files in order to test the
  various combinations

Ideally, instead of having each one of these tests be described a
comment, it would have its own `it` block. Unfortunately right now the
setup would demand running through all the setup steps if we were to do
that, making it very slow.

Instead it is harder to read, but a nice future refactor would let us
pull out the setup in something like a `before(:context)`, allowing us
have proper `it` blocks and change what we need without repetition.

Co-authored-by: Brittany Greaner <brittany.greaner@gsa.gov>
Co-authored-by: Eileen McFarland <eileen.mcfarland@gsa.gov>