Add `Pii::Address` and use it store address in `Idv::AddressController`
Currently `Idv::AddressController` collects an updated address from the user and merges it into `pii_from_doc` in `idv_session`. This results in the address that was read from the doc being overwritten. This approach leads to a few issues:

1. It necessitates a `address_edited` property on `Idv::Session` and custom code to maintain the value of that column. This code also has a tricky bug where it does not track address edits if the same data is submitted twice.
2. It differs from the in-person workflow where the document address and the user address are maintained seperatly for use in double address verification

This commit introduces a new pattern where the PII collected from the user is stored in an immutable object in the `Idv::Session`. This lets us be confident that the PII we have in the session is in fact what the user entered and operate it on such.

This commit makes changes to the session so it will require a follow-up change to read this data once this has been deployed. All of that will need to be followed by a change to stop writing the data using the old approach.

[skip changelog]