Use form object pattern for `Idv::AddressController` submission
While I was working on #10385 I noticed that the `Idv::AddressController` includes a form object but does not implement the Form Object Pattern the way I expected. Specifically the form was only initialized and used to validate input params. As a result a few of the expected features were missing. Namely error messages did not appear on the rendered view. This commit adjusts `Idv::AddressController` to use the pattern as expected.

This commit also adds specs for `Idv::AddressForm` which I noticed were missing despite the form object having some complexity to account for the `address_edited` value in the `FormResponse`.

[skip changelog]