LG-12920: add analytics event for tracking unexpected sdk error (#10368)
* changelog: Internal, Doc Auth, add analytics event for tracking unexpected sdk error
## 🎫 Ticket

[LG-12920](https://cm-jira.usa.gov/browse/LG-12920)

## 🛠 Summary of changes

Implements the `setUnexpectedErrorCallback` method from [Acuant setup instructions (step #5)](https://github.com/Acuant/JavascriptWebSDKV11?tab=readme-ov-file#setup).

We wanted to implement this now in case it gives us useful messages about the black screen issues in [LG-12671](https://cm-jira.usa.gov/browse/LG-12671).

## 📜 Testing Plan

I could not find a way to test this or to trigger it locally. It needs to be an error that happens after SDK load but before initialization. We would like to have a test for it, or at least have a way to trigger it manually, but we don't believe we have a way to do that.