Stop adding updated address to `idv_session.pii_from_doc`
In #10390 we started using the new `idv_session.updated_user_address` value to represent the user address if it is present. This allows us to stop overwriting the user address that was on the document in `pii_from_doc`. We were not able to actually take this step because of the 50/50 state issues related to session updates.

This commit stops overwriting the PII from the document in `pii_from_doc` when a new address is provided in `Idv::AddressController`

[skip changelog]