Fix h1 error issues on review issues page
Previous to this change we did not have this logic and relied on the
`else` case to display the heading for doc auth errors. This meant if we
had both a doc auth error and a selfie error, it would catch the selfie
error first and display the selfie error `h1` instead.

This change adds a new `result_code_invalid?` method to the
`image_upload_response_presenter`.

Passing this through to the front end allows us to use its logic to
determine which `h1` to display in the `document-capture-warning`
component when we have a doc auth error.

Co-authored-by: Brittany Greaner <brittany.greaner@gsa.gov>
Co-authored-by: Eileen McFarland <eileen.mcfarland@gsa.gov>