Fix body copy error issues on review issues page
Previous to this change we did not have this logic and relied on the
`else` case to display the body copy for doc auth errors. This meant if
we had both a doc auth error and a selfie error, it would catch the
selfie error first and display the selfie error body copy instead.

This change:

- updates the error_generator logic to take doc auth errors into account
  - previously we were letting the doc auth errors fall all the way
  through to the `AlertErrorHandler`, that was then handling doc auth,
  pii, and unknown errors
  - we introduced a new `DocAuthErrorHandler` that we could call before
  selfie errors and took much of the logic out of `AlertErrorHandler`
  - we renamed `AlertErrorHandler` to `UnknownErrorHandler` to better
  reflect what it does
- I also needed to update the default error messages to display what we
expect [as represented in the figma](https://www.figma.com/file/YejFLjUrlzZNjjM8McoudH/MVP%3A-Selfie-UX?type=design&node-id=802-5999&mode=design&t=IgxmyFvrKsLZDqZl-0)
- it was tricky getting the attention without barcode case right - this
resulted in some of the changes in the next commit, about the misc other
changes needed for this work.

Co-authored-by: Brittany Greaner <brittany.greaner@gsa.gov>
Co-authored-by: Eileen McFarland <eileen.mcfarland@gsa.gov>
Co-authored-by: Dawei Wang <dawei.wang@gsa.gov>