Remove unused keys from `Idp::Constants::MOCK_IDV_APPLICANT`
The value of  `Idp::Constants::MOCK_IDV_APPLICANT` is returned from `DocAuth::Mock::ResultResponse#pii_from_doc` if no YAML file is provided. This value ultimately gets written to `pii_from_doc` in `Idv::Session`.

The implementation of `DocAuth::Mock::ResultResponse#pii_from_doc` is intended to mirror the implementation of the same method on the LexisNexis TrueID client. This method does not return a hash with a `same_address_as_id`. This commit removes it so the implementation matches.

The same is true of the `phone` key. That is entered by the user downstream of the state ID collection and is not returned by the LexisNexis TrueID client. There is a separate `MOCK_IDV_APPLICANT_WITH_PHONE` constant to cover cases where the phone number is needed.

Finally, this commit adds `issuing_country_code` which is returned by the LexisNexis TrueID client but was not present in the hash.

[skip changelog]