Hoisted :user_can_pass_after_state_id_check?
Turns out there were several useless mocks.

The mocks were a `allows`, not `expects`, so they weren't checking for
anything.

Upon investigation (One at a time, I changed the mock to a `expect`,
watched the test in question fail, then inspected both the test and
the progressive proofer code), the mocked method was really not
supposed to be invoked under the specified circumstances.