Return a `Pii::StateId` object from `pii_from_doc` on the mock proofer
In #10478 I modified the mock proofer to always return the full set of keys that the TrueID client returns from the `#pii_from_doc` method.

This commit continues on the work by introducing `Pii:StateId` to hold the data that is returned from `#pii_from_doc`. It is an immutable struct that holds the data and provides assurances about its shape by ensuring that data is present for every key.

In follow-up commits this data structure will be returned from the TrueID client and used to represent the PII from the document in `Idv::Session`.

[skip changelog]