Return default values for missing fields in `DocAuth::Mock::ResultsResponse#pii_from_doc` (#10478)
The `pii_from_doc` method in `DocAuth::Mock::ResultsResponse` is intended to mock the behavior of the `pii_from_doc` method on the LexisNexis TrueID client. That method uses [`DocAuth::LexisNexis::DocPiiReader`](https://github.com/18F/identity-idp/blob/a2b3cfcf78b127f6d8a7d8b5fbd27141add52159/app/services/doc_auth/lexis_nexis/doc_pii_reader.rb) to construct its response. As a result the response will always contain the same keys.

The `pii_from_doc` method in the mock client returns a set of default PII described by [`Idp::Constants::MOCK_IDV_APPLICANT`](https://github.com/18F/identity-idp/blob/a2b3cfcf78b127f6d8a7d8b5fbd27141add52159/lib/idp/constants.rb#L94) if an image is uploaded. If a user uploads a YAML file, however, the YAML can dictate what PII is returned. This commit makes changes to the behavior when a YAML file is uploaded.

Previously when a YAML file had a `document` key the hash under that `document` key would be returned from `pii_from_doc`. This leads to issues with missing or extra values being present in this hash. This is especially problematic when the YAML file is used for testing since it does not properly simulate the behavior of a deployed environment.

Now, when a `document` key is provided the values from the hash under that `document` key are merged into the default values. This addresses the issue of missing keys. Extraneous keys will be addressed in a follow-up.

### Returning a nil value from `pii_from_doc`

It should be noted that prior to this commit when no `document` key was present the `pii_from_doc` method returned `nil`. This allowed testing error conditions where there would be no PII. For example, if a document cannot be read then `pii_from_doc` is expected to return `nil` since the PII from the document could not be read. This commit preservers that behavior.

### Returning a nil value for specific attributes

There are testing scenarios where it is desirable for a specific PII field to be nil. For example, if we are testing the validation that a zipcode is included we need to simulate a scenario where `pii_from_doc` returns a nil zipcode. Previously those scenarios could be tested by simply not providing a zipcode. Now a default value will be in the place of the absent zipcode. This testing scenario can be covered by explicitly providing a nil value in the YAML file e.g.

```yaml
document:
  zipcode: null
```

[skip changelog]