Return a `Pii::StateId` object from `DocAuth::LexisNexis::Responses::TrueIdResponse#pii_from_doc`
We are working to return immutable structs to represent PII from doc auth vendors that read PII from state IDs. Previous commits (10470, #10488) laid some groundwork for doing this in the mock proofer. This commit takes the step of using the immutable `Pii::StateId` object for PII in the LexisNexis TrueID client.

[skip changelog]