LG-12994: Fix order of errors displayed when there are multiple errors (#10450)
* Test h1 and body copy logic for doc auth + selfie

This change:

- moves some test setup to before all / after all
  - this is so we can reuse the user and other vars we do not change in
  this file and do not care about
  - the one drawback is that currently if you ctrl + c twice out of this
  test, the `after(:all)` block will not run and you will end up with a
  user hanging out in the db that will not be deleted
    - we discussed fixing this in this test, but decided it wasn't a
    good time. If folks agree we can create a ticket to come back to
    this work
    - there is no issue if you let the test finish on its own, or if you
    only ctrl + c once, letting it clean up and run the `after` block
      - pressing it twice skips the `after` block, which is how it ends
      up in this broken state
- adds a lot of test fixture files
  - we needed a lot of different setup files in order to test the
  various combinations

Ideally, instead of having each one of these tests be described a
comment, it would have its own `it` block. Unfortunately right now the
setup would demand running through all the setup steps if we were to do
that, making it very slow.

Instead it is harder to read, but a nice future refactor would let us
pull out the setup in something like a `before(:context)`, allowing us
have proper `it` blocks and change what we need without repetition.

* Fix h1 error issues on review issues page

Previous to this change we did not have this logic and relied on the
`else` case to display the heading for doc auth errors. This meant if we
had both a doc auth error and a selfie error, it would catch the selfie
error first and display the selfie error `h1` instead.

This change adds a new `result_code_invalid?` method to the
`image_upload_response_presenter`.

Passing this through to the front end allows us to use its logic to
determine which `h1` to display in the `document-capture-warning`
component when we have a doc auth error.

* Fix body copy error issues on review issues page

Previous to this change we did not have this logic and relied on the
`else` case to display the body copy for doc auth errors. This meant if
we had both a doc auth error and a selfie error, it would catch the
selfie error first and display the selfie error body copy instead.

This change:

- updates the error_generator logic to take doc auth errors into account
  - previously we were letting the doc auth errors fall all the way
  through to the `AlertErrorHandler`, that was then handling doc auth,
  pii, and unknown errors
  - we introduced a new `DocAuthErrorHandler` that we could call before
  selfie errors and took much of the logic out of `AlertErrorHandler`
  - we renamed `AlertErrorHandler` to `UnknownErrorHandler` to better
  reflect what it does
- I also needed to update the default error messages to display what we
expect [as represented in the figma](https://www.figma.com/file/YejFLjUrlzZNjjM8McoudH/MVP%3A-Selfie-UX?type=design&node-id=802-5999&mode=design&t=IgxmyFvrKsLZDqZl-0)
- it was tricky getting the attention without barcode case right - this
resulted in some of the changes in the next commit, about the misc other
changes needed for this work.

* Misc changes needed to fix up errors displayed

This change:

- updates the mock proofer (and spec) to not delete the attention with
barcode message
  - having the mock proofer delete the attention with barcode error
  message made it impossible to test
- update TrueID specs
  - these began failing with the change to the default liveness general
  error
- fixes an empty space in one of the yml files
  - this was a strange one! [See this Slack comment for why](https://gsa-tts.slack.com/archives/C05HSH9RQ57/p1713286668833529).

* changelog: User-Facing Improvements, Doc Auth, Fix order of errors displayed when there are multiple errors

---------

Co-authored-by: Brittany Greaner <brittany.greaner@gsa.gov>
Co-authored-by: Eileen McFarland <eileen.mcfarland@gsa.gov>
Co-authored-by: Dawei Wang <dawei.wang@gsa.gov>
Co-authored-by: Amir Reavis-Bey <amir.reavis-bey@gsa.gov>