Cleanup `DocPiiReader#read_pii` (#10515)
This commit is in response to feedback on #10498 [ref](https://github.com/18F/identity-idp/pull/10498#discussion_r1578427544).

This method rebuilds the `#read_pii` method so it no longer maintains a map of keys to values and instead reads everything in-line where it can. I am hopeful this makes it a bit more succint and readable.

changelog: Internal, LexisNexis TrueID Response Mixins, The LexisNexis TrueID client read_pii was refactored to improve readability by removing the map that mapped keys in the LexisNexis TrueID response to keys in the eventual pii_from_doc return value and replacing it with a StateID struct constructor call where the keys and values were mapped inline with the exception of the state_id_type value which needs to be independently computed to map it to the appropriate string based on the value in the LexisNexis TrueID response.