Return a `Pii::StateId` object from `idv_session.pii_from_doc`
Previous commits added the `Pii::StateId` data class. This is intended to hold the data that is read from the state ID. This has several advantages over the old approach (holding this data in a Hash):

1. It is immutable so it will error if you try to add data not from the state ID
2. It has defined attributes that are all required so you know what data will be present

This commit takes the next step in integrating this tool which is modifying `idv_session#pii_from_doc` to serialize the PII before it is written to the session and then dersialize it into `Pii::StateId` on read.

[skip changelog]