Refactor `OpenidConnectUserInfoPresenter` to use `AuthnContextResolver`
The `OpenidConnectUserInfoPresenter` is responsible for building the user info hash. This is stored in the identity token and exposed at the user info endpoint as part of the OIDC protocol.

The `OpenidConnectUserInfoPresenter` is reponsible for looking at the parameters of the request to determine which attributes to include in the hash. This can include things like the level of service for identity proofing and the requested scopes.

Prior to this commit the `OpenidConnectUserInfoPresenter` was using the `ServiceProviderIdentity#ial` to determine whether identity proofing was performed. This approach does not work when using multiple vectors of trust determin the authentication context. The `#ial` column represents the level of service that was set when the identity was linked when the SP request was stored in the session. This could change with multiple vectors of trust if the state of the users account changes during the transaction.

This commit starts using the `AuthnContextResolver` to compute the level of service and set the attributes so that it matches the dynamic behavior of the rest of the application with multiple vectors of trust in play.

[skip changelog]