Stop creating `ServiceProviderRequest` records with unused fields
Previous commits retired the following fields on the `ServiceProviderRequest` model:

- ial
- aal
- biometric_comparison_required

These fields were left behind because removing them leads to issues during deploys when old and new versions of the code are running.

The `ServiceProviderRequest` model has default values for these fields. This means that we can stop creating records with these fields. However, we use the hash that is written to redis to create `ServiceProviderRequest`s. This means that if we remove the keyword arguments for these fields we will encounter a `ArgumentError` when we try to create a record on a new host with a hash written by an old host.

[skip changelog]