Add tests for IdentityLinker#process_ial logic
The `IdentityLinker` seems to be setting the following attributes on `ServiceProviderIdentities` for reporting purposes:

- `last_ial2_authenticated_at`
- `last_ial1_authenticated_at`
- `verified_at`

These fields seems to be queried for a number of reports and aren't used by any business logic in the IdP.

This behavior is currently untested. You could remove this entire method and the test suite would still be green. This commit tests the current behavior. The current behavior does not appear to account for a number of edge cases e.g. users who sign in with IALMax or users who are signing into multiple service providers. This commit addresses none of those edge cases because I do not understand the ramifications of addressing those edge cases for our reporting.

[skip changelog]