Modify `OpenidConnect::AuthorizationController#link_identity_to_service_provider` to match pattern in `SamlIdpController`
Both the `OpenidConnect::AuthorizationController` and `SamlIdpController` have logic that ultimately invokes the `IdentityLinker` to link a user to a service provider. This code consumes an `ial` value which is used to set values on the `ServiceProviderIdentity` record that are eventually used for analytics and reporting purposes.

In the `SamlIdpController` the `ial` value is computed using `resolved_authn_contenxt_result`. This means that it considers the SP defaults and the content of the SP request. Eventually this will also include the user context when multiple vectors of trust support is added.

Prior to this commit the `OpenidConnect::AuthorizationController` did not use the `resolved_authn_contenxt_result` and instead computed the value itself in its form object. Its computation ignored SP defaults and would ignore the user context when multiple vectors of trust support is added.

This commit modifies the `OpenidConnect::AuthorizationController1` to match the pattern on `SamlIdpController` to avoid these issues.

[skip changelog]