Refactor `SamlRequestPresenter` into `SamlRequestAttributesPresenter`
The `SamlRequestPresenter` was initialized with the service provider and with the SAML request. It implemented a single public method: `requested_attributes`. This method considered the attributes requested in the SAML request, the service provider defaults, and the IAL context of the request. This was used by the `FederatedProtocols::Saml` class to determine what attributes were requested which ultimately found their way into a column the `identities` table.

Much of the work to consider the IAL context, namely what the service provider requested and service provider defaults, is already handled by the caller. I refactored this class to take the `ial` as an argument instead of determining it from the request object and service provider.

I also moved the `SamlRequestParser` call that found the requested attributes up into the caller and passed that as an argument. That makes the testing approach for this much easier. You no longer need to pass a parsable request object or mock the request parser to test the behavior here.

[skip changelog]