Add and update specs to reflect doc + add ymls
This adds specs for

1. when there are multiple doc auth errors on front and back
2. when there are multiple front doc auth errors
3. when there are multiple back doc auth errors

It also updates the previously existing tests to check for the rate
limit message, as when I was working on the rate limit message change I
found sometimes I had the message showing twice or the wrong message.

So I wanted to make sure that the correct message was showing in those
tests as well.