Simplify rate limit message logic
After discussing this in the error doc, we found that we:

- always want the rate limit message to show
- do not know the logic behind showing one message versus the other, and
would like to opt for the one that is more clear

To that end, I moved the logic for the rate limit message up a level and
out of so many conditionals. This ended up being a bigger change than I
expected as so many tests relied on the place where it was previously
being called, but I think it simplifies the code.