LG-12994: fix messages when there are multiple doc auth errors (#10635)
* Add and update specs to reflect doc + add ymls

This adds specs for

1. when there are multiple doc auth errors on front and back
2. when there are multiple front doc auth errors
3. when there are multiple back doc auth errors

It also updates the previously existing tests to check for the rate
limit message, as when I was working on the rate limit message change I
found sometimes I had the message showing twice or the wrong message.

So I wanted to make sure that the correct message was showing in those
tests as well.

* Restore previous default error message

This reverts a change I made in https://github.com/18F/identity-idp/pull/10450
that turned out to be the wrong change.

* Show subheading whenever IPP and !failedResult

Our previous logic for the subheading was too strict and didn't show
when selfie had errors.

But really we want this h2 to show whever there is a corresponding h2
in the ipp section to balance out the html structure and make it more
accessible.

* Simplify rate limit message logic

After discussing this in the error doc, we found that we:

- always want the rate limit message to show
- do not know the logic behind showing one message versus the other, and
would like to opt for the one that is more clear

To that end, I moved the logic for the rate limit message up a level and
out of so many conditionals. This ended up being a bigger change than I
expected as so many tests relied on the place where it was previously
being called, but I think it simplifies the code.

* Update test helper for mobile ipp selfie flow

Recently this text was changed, and when trying to use the same test
helper as in previous specs, it wouldn't work. This change updates the
helper to be able to navigate the features enabled.

* Rm unusued string from translations

Because we decided to consolidate and only use one of the rate limit
messages, we no longer use this string.

* changelog: User-Facing Improvements, Doc Auth, Fix error messaging for case of multiple errors

---------

Co-authored-by: Amir Reavis-Bey <amir.reavis-bey@gsa.gov>