Update namespaces in AAMVA verification request to match docs
The DLDV documentation PDF we have uses `nc:` to namespace things that are based on NIEM Core and `aa:` to namespace AAMVA extensions to NIEM core.

Our code was using `ns1:` and `ns2:` which made it somewhat hard to follow.

This commit updates template and fixture files to use `aa:` and `nc:` namespaces. It will require testing in the staging environment before deploying.