Add multiple vector of trust support for SAML
We added support for sending multiple vectors of trust using the OIDC interface in #10517. This commit adds the same feature to the SAML interace.

To send multiple vectors using SAML partners can include multiple AuthnContextNodes with a vector in each. This was enabled in https://github.com/18F/saml_idp/pull/100.

[skip changelog]