LG-13038 Add user context to the `​​AuthnContextResolver` (#10517)
The `AuthnContextResolver#resolve` takes the ACR values and VTR param for a request and returns an `Vot::Parser::Result` object. This object has predicate methods such as `aal2?` and `#identity_proofing?` which describe the requirements for the given service provider request. This is necessary because these requirements are a function of the service provider settings that are stored in the database and the request from the service provider.

When using vectors of trust a service provider is able to make a request with multiple valid vectors. Prior to this commit we used the first vector in the list for the service provider request. This commit makes a change to show the following preference:

1. If a SP requests biometric comparison and the user has completed proofing with biometric comparison we select the biometric comparison vector
2. If a SP requests identity proofing without biometric comparison and the user completed proofing we select the vector for identity proofing without biometric comparison
3. If neither of the above is true then we continue to select the first vector.

This change requires passing the current user into the `AuthnContextResolver` initializer so it has access to the user context once the user signs in.

The arrangement allows us to use the following vector combinations to satisfy the following use cases

- `C1.C2.P1.Pb,C1.C2.P1`: Existing proofed users do not need to proof again but unproofed users need to proof with a biometric
- `C1.C2.P1,C1.C2`: Users who have proofed have their attributes shared but unproofed users do not need to go through proofing i.e. IALMax

[skip changelog]