LG-13312 Return an array from `#requested_vtr_authn_context` (#100)
The `#requested_vtr_authn_context` method returns the vectors of trust that appeared in AuthnContextClassRef nodes in the SAML request.

Service providers can send multiple vectors by adding additional AuthnContextClassRef nodes with vectors of trust in them. We initially managed this by selecting the first vector. However, we are interested in adding logic that selects an appropriate vector based on the user context. As a result we need to know all of the vectors that were passed in the SAML request.

This commit modifies the `Request#requested_vtr_authn_context` so that it returns all of the vectors in the SAML request in an array.