Remove `#is_a?` call from VTR in SAML request validator (#10688)
When I was working on the code to parse a VTR in the SAML request validator I copied the code from the OIDC form object. This included a check that the `vtr` is an array. In OIDC that is necessarily because we are directly parsing the result of calling `JSON.parse` on the param. In SAML that is not the case; the vtr is parsed out of the authn request and will be an array or nil. For this reason we can use `#blank?` in the SAML implementation.

[skip changelog]