LG-13424 Fix phone confirmation OTP length message during sign up in auth (#10694)
In #10480 we added an A/B test that made voice OTPs 10-digits for IdV users for some percentage of users. This was to test whether a 10-digit numeric code sees more success than a 6-digit alphanumeric code.

This included a change to the message for the voice OTP since the message includes the length, e.g. "Your 10-digit Login.gov one time code is 1234567890". The translation for the 10-digit code was erroneously applied to 6-digit codes when auth users chose voice calls to confirm phones during sign up. This commit fixes that issue.

[skip changelog]