LG-12555 Use 10-digit voice OTP for all IdV users
We performed an A/B test to determine which of the following so higher rates of OTP confirmation for voice OTPs:

1. 6 character alphanumeric OTPs
2. 10 digit voice OTPs

Analysis of the results determined that 10 digit numeric voice OTPs outperform 6 character alphanumeric voice OTPs.

This commit makes the state where 10 digit voice OTPs are sent during proofing the state for all users requesting voice OTPs during proofing. This commit also removes the code to support the A/B test.

[skip changelog]